task: clean up profile/store code

+++ src/main/kotlin/com/zwolsman/bombastic/controllers/game/payload/CreateGamePayload.kt
@@ -1,7 +1,7 @@
 package com.zwolsman.bombastic.controllers.game.payload
 
 data class CreateGamePayload(
-    val initialBet: Int,
+    val initialBet: Long,
     val bombs: Int,
     val colorId: Int,
 )
