+++ src/main/kotlin/com/zwolsman/bombastic/domain/Tile.kt
@@ -1,5 +1,5 @@
 package com.zwolsman.bombastic.domain
 
 sealed class Tile(val id: Int)
-class Points(id: Int, val amount: Int) : Tile(id)
+class Points(id: Int, val amount: Long) : Tile(id)
 class Bomb(id: Int, val revealedByUser: Boolean) : Tile(id)
